вынесла состояние в верхний компонент из main-screen и offer-screen
